Add functionality to reload default engine config
The code has been updated to include a new function, reloadDefaultConfig, in the Config and Application classes. This new method allows the engine configuration to be reloaded at runtime. Changes have also been made to the Application class to enable this functionality, and the code for the key bindings in main.cpp has been updated to include a command for reloading the engine configuration.