Adjust RendererAPI to interface with MainRenderThread
The RendererAPI has been modified to engage with MainRenderThread instead of Application. This adjustment involves introduction of a new namespace 'threads'. The Application logic has been altered to create and manage these threads. Correspondingly, handling of render context and render API is moved to MainRenderThread. This change aims for better scalability and separation of concerns.