Add thread identification to logging system
Implemented an extension to the logging system that includes thread identification in the log messages. The class `ThreadFormatterFlag` was created to handle this, associating thread names with thread IDs. Modifying `SecondaryThread` and `MainRenderThread` classes allowed individual threads to be named. This enhancement will improve the debugging and tracking capabilities of the logging system.