Refactor VulkanSwapchain and VulkanSwapchainFramebuffers classes
This update refactors the VulkanSwapchain and VulkanSwapchainFramebuffers classes for better code readability and program efficiency. The VulkanSwapchain class has been updated to use a vector of VulkanSwapchainFramebuffer objects, replacing the former Scope<VulkanSwapchainFramebuffers> object. Changes include the removal of undesired methods and properties, and the replacement of several methods with a more streamlined logic flow. This refactor also introduced a new VulkanSwapchainImages class, improving functionality and making the code more comprehensible.