Update rendering and update context structure
In this commit, we have restructured the OpenGL and Vulkan rendering contexts for better support of multithreading. This involved the introduction of new methods to handle frame synchronization and image presentation directly. Simultaneously, a similar approach has been applied to the update context, meaning the update and render loops are now more closely aligned with each other. We have also amended some resource management and function calling to address specific bug fixes in the VulkanSwapchainFramebuffer.